third_party/blink/web_tests/fast/css/html-attr-case-sensitivity-expected.txt

This tests that [attr=value] CSS selectors are case sensitive depending on attr

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS isCaseSensitive("abbr") is true
PASS isCaseSensitive("accept-charset") is false
PASS isCaseSensitive("accept") is false
PASS isCaseSensitive("accesskey") is true
PASS isCaseSensitive("action") is true
PASS isCaseSensitive("align") is false
PASS isCaseSensitive("alink") is false
PASS isCaseSensitive("alt") is true
PASS isCaseSensitive("archive") is true
PASS isCaseSensitive("axis") is false
PASS isCaseSensitive("background") is true
PASS isCaseSensitive("bgcolor") is false
PASS isCaseSensitive("border") is true
PASS isCaseSensitive("cellpadding") is true
PASS isCaseSensitive("cellspacing") is true
PASS isCaseSensitive("char") is true
PASS isCaseSensitive("charoff") is true
PASS isCaseSensitive("charset") is false
PASS isCaseSensitive("checked") is false
PASS isCaseSensitive("cite") is true
PASS isCaseSensitive("class") is true
PASS isCaseSensitive("classid") is true
PASS isCaseSensitive("clear") is false
PASS isCaseSensitive("code") is true
PASS isCaseSensitive("codebase") is true
PASS isCaseSensitive("codetype") is false
PASS isCaseSensitive("color") is false
PASS isCaseSensitive("cols") is true
PASS isCaseSensitive("colspan") is true
PASS isCaseSensitive("compact") is false
PASS isCaseSensitive("content") is true
PASS isCaseSensitive("coords") is true
PASS isCaseSensitive("data") is true
PASS isCaseSensitive("datetime") is true
PASS isCaseSensitive("declare") is false
PASS isCaseSensitive("defer") is false
PASS isCaseSensitive("dir") is false
PASS isCaseSensitive("disabled") is false
PASS isCaseSensitive("enctype") is false
PASS isCaseSensitive("face") is false
PASS isCaseSensitive("for") is true
PASS isCaseSensitive("frame") is false
PASS isCaseSensitive("frameborder") is true
PASS isCaseSensitive("headers") is true
PASS isCaseSensitive("height") is true
PASS isCaseSensitive("href") is true
PASS isCaseSensitive("hreflang") is false
PASS isCaseSensitive("hspace") is true
PASS isCaseSensitive("http-equiv") is false
PASS isCaseSensitive("id") is true
PASS isCaseSensitive("ismap") is true
PASS isCaseSensitive("label") is true
PASS isCaseSensitive("lang") is false
PASS isCaseSensitive("language") is false
PASS isCaseSensitive("link") is false
PASS isCaseSensitive("longdesc") is true
PASS isCaseSensitive("marginheight") is true
PASS isCaseSensitive("marginwidth") is true
PASS isCaseSensitive("maxlength") is true
PASS isCaseSensitive("media") is false
PASS isCaseSensitive("method") is false
PASS isCaseSensitive("multiple") is false
PASS isCaseSensitive("name") is true
PASS isCaseSensitive("nohref") is false
PASS isCaseSensitive("noresize") is false
PASS isCaseSensitive("noshade") is false
PASS isCaseSensitive("nowrap") is false
PASS isCaseSensitive("object") is true
PASS isCaseSensitive("onblur") is true
PASS isCaseSensitive("onchange") is true
PASS isCaseSensitive("onclick") is true
PASS isCaseSensitive("ondblclick") is true
PASS isCaseSensitive("onfocus") is true
PASS isCaseSensitive("onkeydown") is true
PASS isCaseSensitive("onkeypress") is true
PASS isCaseSensitive("onkeyup") is true
PASS isCaseSensitive("onload") is true
PASS isCaseSensitive("onmousedown") is true
PASS isCaseSensitive("onmousemove") is true
PASS isCaseSensitive("onmouseout") is true
PASS isCaseSensitive("onmouseover") is true
PASS isCaseSensitive("onmouseup") is true
PASS isCaseSensitive("onreset") is true
PASS isCaseSensitive("onselect") is true
PASS isCaseSensitive("onsubmit") is true
PASS isCaseSensitive("onunload") is true
PASS isCaseSensitive("profile") is true
PASS isCaseSensitive("prompt") is true
PASS isCaseSensitive("readonly") is false
PASS isCaseSensitive("rel") is false
PASS isCaseSensitive("rev") is false
PASS isCaseSensitive("rows") is true
PASS isCaseSensitive("rowspan") is true
PASS isCaseSensitive("rules") is false
PASS isCaseSensitive("scheme") is true
PASS isCaseSensitive("scope") is false
PASS isCaseSensitive("scrolling") is false
PASS isCaseSensitive("selected") is false
PASS isCaseSensitive("shape") is false
PASS isCaseSensitive("size") is true
PASS isCaseSensitive("span") is true
PASS isCaseSensitive("src") is true
PASS isCaseSensitive("standby") is true
PASS isCaseSensitive("start") is true
PASS isCaseSensitive("style") is true
PASS isCaseSensitive("summary") is true
PASS isCaseSensitive("tabindex") is true
PASS isCaseSensitive("target") is false
PASS isCaseSensitive("text") is false
PASS isCaseSensitive("title") is true
PASS isCaseSensitive("type") is false
PASS isCaseSensitive("usemap") is true
PASS isCaseSensitive("valign") is false
PASS isCaseSensitive("value") is true
PASS isCaseSensitive("valuetype") is false
PASS isCaseSensitive("version") is true
PASS isCaseSensitive("vlink") is false
PASS isCaseSensitive("vspace") is true
PASS isCaseSensitive("width") is true
PASS isCaseSensitive("foobar") is true
PASS successfullyParsed is true

TEST COMPLETE
